User microservice

+++ backend/user-service/controllers/user-controller.ts
@@ -4,6 +4,20 @@ import bcrypt from "bcrypt";
 import db from "../models/user-model";
 import HttpStatusCode from "../libs/enums/HttpStatusCode";
 
+async function health(req: Request, res: Response) {
+  try {
+    await db.checkDatabase();
+    return res.json({ 
+      message: "User microservice is working." 
+    });
+  } catch (err) {
+    console.log(err);
+    return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR.valueOf()).json({
+      message: "Internal User microservice internal error."
+    });
+  }
+}
+
 async function registerUser(req: Request, res: Response) {
   const { email, password, name, major, role } = req.body;
 
@@ -34,7 +48,10 @@ async function registerUser(req: Request, res: Response) {
             hash,
             role
           );
-          return res.json({ uid });
+          return res.json({
+            uid: uid,
+            message: "User registered successfully."
+          });
         } catch (err) {
           console.log(err);
           return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR.valueOf()).json({
@@ -49,7 +66,7 @@ async function registerUser(req: Request, res: Response) {
   } catch (err) {
     console.log(err);
     return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR.valueOf()).json({
-      message: "Undefined error creating users.",
+      message: "Internal server error creating users.",
     });
   }
 }
@@ -79,7 +96,7 @@ async function loginUser(req: Request, res: Response) {
           if (!jwtSecretKey) {
             console.error("JWT secret key is not defined.");
             return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR.valueOf()).json({
-              message: "Internal server error.",
+              message: "Internal server error cannot authenticate user logging in",
             });
           }
 
@@ -104,7 +121,7 @@ async function loginUser(req: Request, res: Response) {
       })
       .catch((err) => {
         console.log(err);
-        return res.send({ message: "Error checking password." });
+        return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR.valueOf()).send({ message: "Internal server error checking password." });
       });
   }
 }
@@ -113,7 +130,7 @@ async function getUserInfo(req: Request, res: Response) {
   const queryUidString = req.query.uid;
   console.log(queryUidString);
   if (typeof queryUidString !== 'string') {
-    return res.status(HttpStatusCode.BAD_REQUEST.valueOf()).json({ message: 'Invalid uid.' });
+    return res.status(HttpStatusCode.BAD_REQUEST.valueOf()).json({ message: "Invalid uid." });
   }
 
   try {
@@ -122,59 +139,38 @@ async function getUserInfo(req: Request, res: Response) {
     const userIdSearch = await db.getUserByUserId(uid);
     if (userIdSearch.rows.length == 0) {
       console.log("User does not exist.");
-      return res.status(HttpStatusCode.BAD_REQUEST.valueOf()).json({
+      return res.status(HttpStatusCode.NOT_FOUND.valueOf()).json({
         message: "User does not exist.",
       });
     } else if (userIdSearch.rows.length > 0) {
-      const user = userIdSearch.rows[0];
+      const userRow = userIdSearch.rows[0];
+      const user = {
+        uid: userRow.uid,
+        name: userRow.name,
+        email: userRow.email,
+        major: userRow.major,
+        avatarUrl: userRow.avatarUrl,
+        role: userRow.role,
+        bio: userRow.bio,
+      }
       return res.json(user);
     }
   } catch (err) {
     console.log(err);
     return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR.valueOf()).json({
-      message: "Error getting user by uid.",
+      message: "Internal server error getting user by uid.",
     });
-  }
-}
-
-async function getUserByEmail(req: Request, res: Response) {
-  const { email } = req.body;
-  try {
-    const emailSearch = await db.getUserByEmail(email);
-    if (emailSearch.rows.length == 0) {
-      console.log("User does not exist.");
-      return res.status(HttpStatusCode.BAD_REQUEST.valueOf()).json({
-        message: "User does not exist.",
-      });
-    } else if (emailSearch.rows.length > 0) {
-      const user = emailSearch.rows[0];
-      return res.json(user);
-    }
-  } catch (err) {
-    console.log(err);
-    return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR.valueOf()).json({
-      message: "Error getting user by email.",
-    });
-  }
-}
-
-async function getAllUsers(req: Request, res: Response) {
-  try {
-    const allUsers = await db.getAllUsers();
-    return res.json(allUsers);
-  } catch (err) {
-    console.log(err);
-    return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR.valueOf()).send({ message: "Error getting all users." });
   }
 }
 
 async function updateUserPassword(req: Request, res: Response) {
   const { uid, old_password, new_password } = req.body;
+
   try {
     const userIdSearch = await db.getUserByUserId(uid);
     if (userIdSearch.rows.length == 0) {
       console.log("User does not exist.");
-      return res.status(HttpStatusCode.FORBIDDEN.valueOf()).json({
+      return res.status(HttpStatusCode.NOT_FOUND.valueOf()).json({
         message: "User does not exist.",
       });
     } else if (userIdSearch.rows.length > 0) {
@@ -199,14 +195,14 @@ async function updateUserPassword(req: Request, res: Response) {
                   });
                 } catch (err) {
                   return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR.valueOf()).json({
-                    message: "Failed to update user password.",
+                    message: "Internal server error updating user password.",
                   });
                 }
               })
               .catch((err) => {
                 console.log(err);
                 return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR.valueOf()).send({
-                  message: "Error crypting password.",
+                  message: "Internal server error updating user password.",
                 });
               });
           }
@@ -214,14 +210,14 @@ async function updateUserPassword(req: Request, res: Response) {
         .catch((err) => {
           console.log(err);
           return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR.valueOf()).send({
-            message: "Error checking password.",
+            message: "Internal server error updating user password.",
           });
         });
     }
   } catch (err) {
     console.log(err);
     return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR.valueOf()).send({
-      message: "Error getting user by uid.",
+      message: "Internal server error updating user password.",
     });
   }
 }
@@ -229,29 +225,33 @@ async function updateUserPassword(req: Request, res: Response) {
 async function updateUserInfo(req: Request, res: Response) {
   const queryUidString = req.query.uid;
   if (typeof queryUidString !== 'string') {
-    return res.status(HttpStatusCode.BAD_REQUEST.valueOf()).json({ message: 'Invalid uid.' });
+    return res.status(HttpStatusCode.BAD_REQUEST.valueOf()).json({ message: "Invalid uid." });
   }
   const uid = parseInt(queryUidString);
   const updateFields = req.body;
 
   try {
     if (Object.keys(updateFields).length === 0) {
-      return res.status(HttpStatusCode.BAD_REQUEST.valueOf()).json({ message: 'No fields provided for update.' });
+      return res.status(HttpStatusCode.BAD_REQUEST.valueOf()).json({ message: "No fields provided for update." });
     }
 
     await db.updateUserInfo(uid, updateFields);
 
     return res.json({
-      message: 'User info updated.',
+      message: "User information updated.",
     });
   } catch (err) {
-    console.error('Error updating user info:', err);
-    return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR.valueOf()).json({ message: 'Failed to update user info.' });
+    console.error("Error updating user info:", err);
+    return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR.valueOf()).json({ message: "Internal server error updating user info." });
   }
 }
 
 async function deleteUser(req: Request, res: Response) {
-  const { uid } = req.body;
+  const queryUidString = req.query.uid;
+  if (typeof queryUidString !== 'string') {
+    return res.status(HttpStatusCode.BAD_REQUEST.valueOf()).json({ message: "Invalid uid." });
+  }
+  const uid = parseInt(queryUidString);
   try {
     const result = await db.deleteUser(uid);
     console.log(result);
@@ -261,27 +261,18 @@ async function deleteUser(req: Request, res: Response) {
     });
   } catch (err) {
     console.log(err);
-    return res.send({
-      message: "Undefined error deleting account.",
+    return res.status(HttpStatusCode.INTERNAL_SERVER_ERROR.valueOf()).send({
+      message: "Internal server error deleting account.",
     });
   }
 }
 
-async function clearCookie(req: Request, res: Response) {
-  res.clearCookie("token");
-  return res.send({
-    message: "Cleared user cookie",
-  });
-}
-
 export default {
+  health,
   registerUser,
   loginUser,
   getUserInfo,
-  getUserByEmail,
-  getAllUsers,
   updateUserPassword,
   updateUserInfo,
   deleteUser,
-  clearCookie,
 };
